backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
rasterio==1.3.9
xarray==2023.12.0
netcdf4==1.6.5
numpy==1.26.2
pillow==10.1.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic-extra-types==2.3.0

